Change noto font weight numbering to use 'pseudo-css' values.
Formerly noto_fonts identified the expected os/2 weight values in
the fonts in its WEIGHT table.  Currently we still encode suboptimal
values to work around legacy systems that autobolded weights they
considered too low, by assigning 250 as the lowest weight.  This means
the weight values can't distinguish between ExtraLight and Thin.

For CSS we were reusing these values, and that introduces yet other
problems.  CSS currently doesn't handle weights that are not multiples
of 100 between 100 and 900.  Not only these weights failed, but we
also have a 'DemiLight' value for Noto Sans CJK between Light and
Regular that was assigned an os/2 weight value of 350.

This commit changes the semantics of the WEIGHT values to be
'pseudo-css', in that they are allowed to take values below 250 but
are not limited to multiples of 100.  This lets us provide a mapping
from our font weight names to numeric weights that correctly represents
the ordering of these weights, and mostly corresponds to the expected
internal weight values.

To accomodate this we needed to change tests that relied on the
encoded weights matching the os/2 weight values.  These tests now
implement the restriction we expect, i.e. that values under 250
are pinned to 250.

The other use, for generating css files for the website, was already
broken in a few ways.  It duplicated urls for the monospace and
non-monospace cjk sans under the same family/weight/slope, and
encoded weights that were not multiples of 100, with the result that
conformant browsers (i.e. chrome) ignored these weights and treated
them as unspecified.  The change here was to encode these under
separate family names, by appending 'Mono' to monospace variants of
the CJK, and by appending the numeric weight value to the family
names of fonts that had non-standard weights and resetting its
value to 700 (Bold).  This allows css to reference these fonts, in
the first case as a standard family name with its own range of weights,
and in the second case as a font with a fixed weight-- defining it
as having a weight of 700 prevents Chrome from autobolding it for
weights 600 and above, which it otherwise would do.

A third use for the weight values is to link font instances to the
samples generated for them.  These values are essentially strings and
represent no other semantic, so are unaffected by this change.