Support legacy Android sequences.
Since Android shipped with support for these sequences, they want to
keep them around even though they're no longer in the unicode emoji
data.

This hard-codes information about the sequences into unicode_data
in the form of text, and changes some code to use them by default.

_LEGACY_ANDROID_SEQUENCES takes the form of the emoji sequence data
and is parsed by the same code, except reworked slightly so that
the text to be parsed can be provided as a list of lines as well
as a filename.

_LEGACY_ANDROID_ORDER is a list of sequences.  Each sequence starting
with '-' is a key, the following are sequences to be added after
the key.  This is used to insert these sequences into the order
data returned by parsing the emoji-test data file.  For example,
'men wrestling' sequences with skin color are inserted after the
'men wrestling' sequence.  This keeps the patch out of the
unicode-supplied data file.

The order data is patched using two functions, one to define a map
of patch data, one to apply it, because who knows but we'll have to
patch more stuff in the future and this seemed cleaner, somehow.

Ideally of course we'd provide a way to get the unpatched data, but
for now the android requirements are driving this.