Fix Adlam sample.
Current libraries don't know Adlam is RTL, and there's no easy way to
fix this.  Work around the issue for now by inserting an RLO in front of
the adlam sample text.

In the course of exploring this I understood better what the flags to
pango layout actually do.  Wrote comments explaining this in the
text image generation code.