Update tool to promote phase3 fonts.
Two main changes:

1) Previously it generated version info by looking at the repo of all of
the source fonts.  This means the fonts have to be in the noto-fonts
repo and selecting the right set of fonts needs to be done on the command
line that invokes the tool.  I'd forgotten this and put the fonts in a
separate directory, at which point the tool could no longer identify
the repo.  This lets you override the check and just stipulate that the
fonts came from (the current version of) the repo.

2) The version wasn't getting updated correctly, in particular phase2
font versions were retaining a major version number of 1.  Must not have
tested this case.  So the logic was changed a bit to identify fonts that
appear to have phase 2 version numbers and ensure they get handled
correctly.