Quiet some git utilities.
Some conditions cause git to print messages to stdout.  This makes
these a bit quieter.

git symbolic-ref complains about detached HEAD when asked what it
points to, since then it's not a symbolic ref.  We now capture stdout
and suppress it, since all we care about is the error.

git_is_clean used to always dump info about unstaged or uncommitted
changes to stdout.  We not add a flag to suppress this, and also
capture the output and truncate it when we don't suppress.  For example,
if we don't check out a directory, all the files that would have been
in that directory look like unstaged deletions-- sometimes hundreds of
files.  We really don't need all this info, just enough to get the
idea, so we truncate the output to 20 lines.