Add a 'joining' sample for Adlam.
This is simply the und-Adlam_chars.txt with the spaces removed and an
RLO in front to force RTL since some libsdon't know about Adlam yet.

Potentially the sample selection code in the website data generation
tooling could know to choose this sample for the joined adlam and the
other for the unjoined, but we don't do that yet, 'text' samples are
always preferred to 'chars' samples.