Improve noto configuration and fontconfig handling.
Notoconfig exists so we can locate other noto repos without reference
to our current location.  It turns out our current location is also
a bit variable depending on how we installed and how we are invoking
the code.

This is a problem for nototools itself because setup.py install
doesn't add all the files some tools need, in particular the sample
texts are not part of the installation.  If we try to locate them
relative to the installed python code, we fail. One fix is to locate
them via the path in the noto config file.

It turns out as well that our fonts.conf (used to restrict cairo to
find only noto fonts) has a similar problem.  This was also assuming
the noto fonts are available at paths relative to the current code
(create_image.py).  This argues again for requiring the configuration
file.  Writing the fonts.conf means we can use the paths we know
from the configuration rather than depending on relative paths and
setting XDG variables.

This change no longer defaults the config value of noto_tools, instead
we require the config file to exist if we ask for it.  It records
the location of the config file for ease of debugging.

It also removes the fonts.conf file and instead writes it when
create_image first needs it, using the locations of the other noto
directories it finds from the noto config file.  The new location for
our custom fonts.conf and the fontconfig cache directory is the root of
the nototools repo rather than under its 'nototools' subdirectory.  This
seemed like better places for them.