Add utility to sync noto font repos to release tags.
The tool to generate the get/noto data by default expects the font
repos to be at a tagged release.  It uses this information to generate
README documents for the download packages.  This enables development
teams that want to bundle our fonts to trace the packages back to
commits in the respective repos.

To simplify and automate this process its helpful to have a tool that
gets the repos to an acceptable state.  By default it checks out the
latest tagged release in each repo.  It is also possible to name a
release tag for any of the repos.  Finally, one can also check out
the master of each repo.