First pass at some Docker tooling.
There is lots of room for refinement here.

This creates a docker subdir with three image configurations.

'toolbase' sets up the libraries most of noto requires, including
harfbuzz, cairo, fontconfig, etc.  It is based on Python 2 so is
much larger than we actually require.

'notobase' pulls in the noto repos, truncating them to a recent
tagged release in a futile attempt to save some space.  Actually
though the fonts make these quite large.  Besides these repos
it also pulls in afdko.  Since some tools expect en-US locale
it also sets that up.  'notobase' is based on 'toolbase'.

'make_getnoto' is based on 'notobase' and is an implementation of
the build process for the data behind the get/noto site. It sets up
the repos at the latest tagged build, generates optipng-compressed
emoji images (typically for font builds we use zopfli, but it is
much slower), builds the website samples, and copies the data to
/tmp/website2.  To use this you need to map this directory in the
docker filesystem to one on your host before running the image.

The images referred to by the dockerfiles are (currently) not
named after the directories, unfortunately.  That would make
setting these up easier.