Add docker image file for noto development.
The existing 'noto' docker files create a noto environment using
the upstream repos under googlei18n.  They're not convenient for
development as the intent is to use the tools but not change the
repos themselves.

This does some of the same setup (e.g. of afdko)  but expects the
noto repos to live on the host and be mapped into the docker
environment when you run the image using -v.

Right now running the image doesn't automatically setup noto for
use.  In the container you should run
/usr/local/share/noto/setup_nototools.sh, this makes the nototools
module visible and installs its requirements.