Support both DemiLight and SemiLight names.
CJK uses DemiLight, but the rest of Noto uses SemiLight for the same
weight.

This means we can no longer map 1-1 from weight to name, but we are
no longer showing those in the get/noto UI page so the only client
of that mapping was removed.  So we remove the mapping as well.