Two tweaks to font ids.
NotoMono is renamed NotoSansMono in phase 3.  The first got family id
'mono-mono', and the second 'sans-mono-mono'.  The second 'mono' in the
latter id clutters up name generation so we remove it.

The family_ids are used for grouping fonts for bundling/display on the
get/noto site, while the wws_ids are used to bundle fonts for naming
purposes.  The semantics are a bit different, wws_ids identify fonts
in the same family whose names only differ in the wws style portion.
This fixes a bug in the wws_id for cjk fonts, it was not distinguishing
between the cjk mono and standard names.

Also adds a few comments.