In phase 3, always include 'Regular' in the full name.
We'd been inconsistent about whether the regular style of the font
would include 'Regular' in the full name.  We had decided to omit
it except for CJK, but Glyphs includes it and the toolchain decided
to follow Glyphs' lead, so now we do too.

This is a bit iffy since it both changes how the data file is
generated (to always request 'Regular' be added) and overrides the
data file (to always add 'Regular' even if not requested for phase 3).