Shorten long script names to fit family name length limit.
Some script names are particularly long and it is not possible to
fit them plus 'noto serif' plus style variants into 32 chars.  So
This adds a table that maps names longer than 10 chars to mangled
names that are 10 chars or less, but hopefully still recognizable.
If the code determines that a family name is too long to fit the
requested non-standard styles it will try the shorter names as
well as shorter style names.

These are for compatibility with limits on old systems and style
linking.  Newer systems should use the preferred name fields anyway
so these names should rarely be seen.

These short names are just made up, no attempt was made to determine
if the chosen abbreviation makes sense to people familiar with these
scripts.