Add flag to force consideration of all wws styles.
Normally, the tool takes font names on the command line, groups them
into families, and records the style variations found for each family.
It then determines whether each family requires preferred names and/or
family name shortening based on this set.

There are two problems with this.  One is we might not have all the
planned style names available-- we can deal with that by synthesizing
names.  The other is that we might add more style names over time,
leading to different kinds of abbreviations for the same (preferred)
family.

Since there isn't a really strong need to save space, or to keep the
family name legible, another option is to just plan for the worst case
from the vary beginning and choose behavior that reflects the presence
of all wws styles (9 weights, 4 condensed levels, italic) in the family.
That is what the new flag does-- it ensures that the longest set of
style names is considered.

Note that this doesn't anticipate possible introduction of unplanned
Mono, Display, or UI variants, which would also further reduce the
space available for the family name.  These are relatively rare though
and would make the family name pretty much unreadable for most all
fonts if it were included.  On the off chance that some users will
see the family name in UIs on old/misbehaving systems, we take the
risk that the family names won't compress similarly in order to get
a bit more readability.