Allow specifying '-' as the info file to write to stdout.
Normally if an info file is not specified data gets written to
stdout, but if a phase is passed (-p) this defaults the info file
to the appropriate name, and some operations require the phase be
passed.  To allow those operations to write to stdout '-' is
now allowed as an info file value (which overrides the default name
triggered by -p).