Clean up script name handling a bit.
We shouldn't be defaulting four-letter script names to codes if they
don't match as names, this should just be an error.  Fix the call
that was doing this and raise the error.

Also add a comment in unicode_data to clarify the distinction between
iso script codes and those assigned by unicode to codepoints.  We try
to maintain this distinction in the data but the apis don't always make
this clear.