Some tooling changes for phase3 releases.
autofix_for_phase3 is a hack.  Originally it did some checks and
swatted the version numbers, because we didn't do that in the source.
But doing as much as possible in the source is better, so the version
update code is not so much needed now.  But it's still there, along
with other stuff we don't need.  Eventually it should go away.  These
changes mostly make it easier to run from docker, on a predefined list
of fonts.  The other tool just generates a list of fonts from a shorter
list of scripts to release and a source directory root.