Update udhr tooling a bit more.
The code that checked commits matching a tool-update message was
raising exceptions if a file in the directory was not under version
control.  This was causing problems so I relaxed that restriction.
I also log some information to stderr.

The udhr tool was one flag to signal whether to protect files that
were in git and also whether to stage a commit.  In docker, we want
to do the first check, but not stage a commit, so I split the flag.