Support sample variants.
We generate a list of possible samples for a given bcp47 language
code, ordering by type.  Formerly we just matched the code and types
directly.  This now matches the code plus any variants and the type.
This allows us to have multiple samples of the same type for a given
bcp47 code, and so support fonts that support the same script but
which have different character repertoires, as the symbols fonts do.