Update cldr data tool.
The new cldr data uses standard 8-digit sequences for '\U'-escaped
codepoints, which formerly we didn't actually use.

The new data also has some anomalies in the likelySubtags data that
broke some of our assumptions (that each lang used in a region had
an entry in likelySubtags), so patch around it for now.