Delete obsolete tools.
emoji_ordering used to generate the emoji ordering data from random data
files in html or csv format, but for some time now we've generated this
from official unicode data files, so we no longer need it.

generate_emoji_name_data was updated and moved to noto-emoji, which was
a better home for it.  The version here is obsolete and should not be
used.