Use auxiliary exemplar data in char sample generation.
Suggested by Jungshik.

The risk here is that the exemplar file might now contain characters
not supported by the fonts that would use them, whereas before the
'standard' exemplars were supported.  However, this does not seem
to have caused there to be no samples for any current font, at least.
This may be because the font supported the additional characters,
or because the UDHR or custom sample text was used for that font
already anyway so the exemplars were never considered.