Add options to create_image.
Adds maxheight and horiz_margin options.

Maxheight is to truncate the text at a given height or number of lines.
The version of pango/cairo we're using doesn't support this so we emulate
it.  Negative values are the number of lines, positive values are the number
of pixels (whole lines).  This is used so that we can limit height of the
generated images at large point sizes (used for Display fonts).

horiz_margin is to allow for italics at large point sizes that extend to
the left of the origin or the right of the advance.  Rather than compute
this automatically (which would lead to varying margins for the samples on
a single page) this is left to the caller to specify.  We also report when
the pixel bounds overflows the margin, so it's possible to determine
when a margin isn't sufficient.  This was a problem for Display Italic
fonts.