Rename Phags-Pa to Phags-pa.
Special-case this name, not only because of the hyphen.  The unicode
data (which provides names in comments, and which we use in preference
to CLDR as these names are somewhat more stable and shorter) uses
underscore and capitalizes the second P.

There was some debate about whether we should use the hyphen.  I think
as long as it's kept out of the postscript name it should be ok...