Support stretch, larger display sizes, and lineheight/margin.
(wasn't able to split hunks finely enough so gave up)

With the addition of condensed/extracondensed etc.,  we now want to
sort by stretch, use stretch in file names, etc.  This adds all stretch
names/abbreviations and handles stretch.

This also supports a larger size for display fonts, as they are intended
to be used at larger sizes.

Because of the large size we need a limit on the vertical height of the
sample, so it's limited to two lines using the (new) option in
create_image.  The large size with italic now causes the glyphs at the
start of some lines to have negative side bearings, so we add a margin
option to increase padding.  It turns out that some of the normal size
fonts had this issue too, so a smaller margin is added for those as well.