Report only non-UI fonts, and check the font count.
Formerly we were counting both UI and non-UI fonts, as the zips contain
both.  This however was being presented on the website as the number of
'styles' and so for fonts with UI variants this was twice as large as
the number of styles we were displaying.  It is better to just count
the styles we show, so we are not counting the UI fonts.

It was discovered in this process that some fonts were unintentionally
present.  There were 4 LGC UI fonts from phase 2 (this were the same
as the non-UI fonts except for the font naming, it was a workaround for
some fallback behavior issues that we no longer need), and there were
3 condensed variants of Bengali that were part of a proposal that was
not actually final and approved for release.  So in order to help detect
this kind of issue we validate against expected counts (this is how the
Bengali error was found).