Update unicode_data to deal with changes in emoji-data.txt.
1) There's a new property and a removed property, define the new
property (different name scheme from all the others) so we can
parse it.  Remove old property to clean up.  Hope no one uses it.

2) The emoji group data now includes the color swatches in the
people and smiley's group.  Previously we patched in an additional
group for them.  This resulted in having two locations for the
same sequences, and the data checks complained.  Went with the
new standard location (they also have hair modifiers in a group),
though most likely for our purposes (showing emoji to users) these
don't need to be grouped as they are but either hidden or tucked
away (as formerly) in their own corner.